mv availabilityzones to compute extensions
