+++ openstack/compute/v2/extensions/availabilityzones/doc.go
@@ -1,0 +1,3 @@
+// Package availabilityzones provides information and interaction with Availability zones
+// extension that works with the OpenStack Compute service.
+package availabilityzones
