+++ openstack/compute/v2/extensions/availabilityzones/testing/doc.go
@@ -1,0 +1,1 @@
+package testing
