+++ openstack/compute/v2/extensions/availabilityzones/urls.go
@@ -1,3 +1,5 @@
+// Package availabilityzones provides information and interaction with Availability zones
+// extension that works with the OpenStack Compute service.
 package availabilityzones
 
 import "github.com/gophercloud/gophercloud"
